Extracted out CrlServer from CertificateRevocationListNodeTests to be inline with ENT
