+++ node/src/integration-test/kotlin/net/corda/node/amqp/CertificateRevocationListNodeTests.kt
@@ -8,11 +8,6 @@ import net.corda.core.internal.div
 import net.corda.core.internal.times
 import net.corda.core.toFuture
 import net.corda.core.utilities.NetworkHostAndPort
-import net.corda.core.utilities.days
-import net.corda.core.utilities.minutes
-import net.corda.core.utilities.seconds
-import net.corda.coretesting.internal.DEV_INTERMEDIATE_CA
-import net.corda.coretesting.internal.DEV_ROOT_CA
 import net.corda.coretesting.internal.rigorousMock
 import net.corda.coretesting.internal.stubs.CertificateStoreStubs
 import net.corda.node.services.config.NodeConfiguration
@@ -33,42 +28,25 @@ import net.corda.testing.core.BOB_NAME
 import net.corda.testing.core.CHARLIE_NAME
 import net.corda.testing.core.MAX_MESSAGE_SIZE
 import net.corda.testing.driver.internal.incrementalPortAllocation
+import net.corda.testing.node.internal.network.CrlServer
+import net.corda.testing.node.internal.network.CrlServer.Companion.FORBIDDEN_CRL
+import net.corda.testing.node.internal.network.CrlServer.Companion.withCrlDistPoint
 import org.apache.activemq.artemis.api.core.RoutingType
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 import org.bouncycastle.asn1.x500.X500Name
 import org.bouncycastle.asn1.x509.*
-import org.bouncycastle.cert.jcajce.JcaX509CRLConverter
-import org.bouncycastle.cert.jcajce.JcaX509ExtensionUtils
-import org.bouncycastle.cert.jcajce.JcaX509v2CRLBuilder
 import org.bouncycastle.jce.provider.BouncyCastleProvider
-import org.bouncycastle.operator.jcajce.JcaContentSignerBuilder
-import org.eclipse.jetty.server.Server
-import org.eclipse.jetty.server.ServerConnector
-import org.eclipse.jetty.server.handler.HandlerCollection
-import org.eclipse.jetty.servlet.ServletContextHandler
-import org.eclipse.jetty.servlet.ServletHolder
-import org.glassfish.jersey.server.ResourceConfig
-import org.glassfish.jersey.servlet.ServletContainer
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.TemporaryFolder
-import java.io.Closeable
 import java.math.BigInteger
-import java.net.InetSocketAddress
-import java.security.KeyPair
-import java.security.PrivateKey
-import java.security.cert.X509CRL
 import java.security.cert.X509Certificate
 import java.time.Duration
 import java.util.*
 import java.util.concurrent.atomic.AtomicInteger
-import javax.ws.rs.GET
-import javax.ws.rs.Path
-import javax.ws.rs.Produces
-import javax.ws.rs.core.Response
 import kotlin.test.assertEquals
 
 @Suppress("LongParameterList")
@@ -76,9 +54,6 @@ class CertificateRevocationListNodeTests {
     @Rule
     @JvmField
     val temporaryFolder = TemporaryFolder()
-
-    private val ROOT_CA = DEV_ROOT_CA
-    private lateinit var INTERMEDIATE_CA: CertificateAndKeyPair
 
     private val portAllocation = incrementalPortAllocation()
     private val serverPort = portAllocation.nextPort()
@@ -93,8 +68,6 @@ class CertificateRevocationListNodeTests {
     private abstract class AbstractNodeConfiguration : NodeConfiguration
 
     companion object {
-        const val FORBIDDEN_CRL = "forbidden.crl"
-
         private val unreachableIpCounter = AtomicInteger(1)
 
         private val crlTimeout = Duration.ofSeconds(System.getProperty("com.sun.security.crl.timeout").toLong())
@@ -107,45 +80,14 @@ class CertificateRevocationListNodeTests {
             check(unreachableIpCounter.get() != 255)
             return "10.255.255.${unreachableIpCounter.getAndIncrement()}"
         }
-
-        private fun createRevocationList(clrServer: CrlServer,
-                                         signatureAlgorithm: String,
-                                         caCertificate: X509Certificate,
-                                         caPrivateKey: PrivateKey,
-                                         endpoint: String,
-                                         indirect: Boolean,
-                                         serialNumbers: List<BigInteger>): X509CRL {
-            println("Generating CRL for $endpoint")
-            val builder = JcaX509v2CRLBuilder(caCertificate.subjectX500Principal, Date(System.currentTimeMillis() - 1.minutes.toMillis()))
-            val extensionUtils = JcaX509ExtensionUtils()
-            builder.addExtension(Extension.authorityKeyIdentifier,
-                    false, extensionUtils.createAuthorityKeyIdentifier(caCertificate))
-            val issuingDistPointName = GeneralName(
-                    GeneralName.uniformResourceIdentifier,
-                    "http://${clrServer.hostAndPort.host}:${clrServer.hostAndPort.port}/crl/$endpoint")
-            // This is required and needs to match the certificate settings with respect to being indirect
-            val issuingDistPoint = IssuingDistributionPoint(DistributionPointName(GeneralNames(issuingDistPointName)), indirect, false)
-            builder.addExtension(Extension.issuingDistributionPoint, true, issuingDistPoint)
-            builder.setNextUpdate(Date(System.currentTimeMillis() + 1.seconds.toMillis()))
-            serialNumbers.forEach {
-                builder.addCRLEntry(it, Date(System.currentTimeMillis() - 10.minutes.toMillis()), ReasonFlags.certificateHold)
-            }
-            val signer = JcaContentSignerBuilder(signatureAlgorithm).setProvider(Crypto.findProvider("BC")).build(caPrivateKey)
-            return JcaX509CRLConverter().setProvider(Crypto.findProvider("BC")).getCRL(builder.build(signer))
-        }
     }
 
     @Before
     fun setUp() {
         // Do not use Security.addProvider(BouncyCastleProvider()) to avoid EdDSA signature disruption in other tests.
         Crypto.findProvider(BouncyCastleProvider.PROVIDER_NAME)
-        crlServer = CrlServer(NetworkHostAndPort("localhost", 0))
+        crlServer = CrlServer(NetworkHostAndPort("localhost", 0), revokedNodeCerts, revokedIntermediateCerts)
         crlServer.start()
-        INTERMEDIATE_CA = CertificateAndKeyPair(replaceCrlDistPointCaCertificate(
-                DEV_INTERMEDIATE_CA.certificate,
-                CertificateType.INTERMEDIATE_CA,
-                ROOT_CA.keyPair,
-                "http://${crlServer.hostAndPort}/crl/intermediate.crl"), DEV_INTERMEDIATE_CA.keyPair)
     }
 
     @After
@@ -285,25 +227,21 @@ class CertificateRevocationListNodeTests {
 	fun `verify CRL algorithms`() {
         val emptyCrl = "empty.crl"
 
-        val crl = createRevocationList(
-                crlServer,
+        val crl = crlServer.createRevocationList(
                 "SHA256withECDSA",
-                ROOT_CA.certificate,
-                ROOT_CA.keyPair.private,
+                crlServer.rootCa,
                 emptyCrl,
                 true,
                 emptyList()
         )
         // This should pass.
-        crl.verify(ROOT_CA.keyPair.public)
+        crl.verify(crlServer.rootCa.keyPair.public)
 
         // Try changing the algorithm to EC will fail.
         assertThatIllegalArgumentException().isThrownBy {
-            createRevocationList(
-                    crlServer,
+            crlServer.createRevocationList(
                     "EC",
-                    ROOT_CA.certificate,
-                    ROOT_CA.keyPair.private,
+                    crlServer.rootCa,
                     emptyCrl,
                     true,
                     emptyList()
@@ -412,7 +350,7 @@ class CertificateRevocationListNodeTests {
             doReturn(crlCheckSoftFail).whenever(it).crlCheckSoftFail
         }
         clientConfig.configureWithDevSSLCertificate()
-        val nodeCert = (signingCertificateStore to p2pSslConfiguration).recreateNodeCaAndTlsCertificates(nodeCrlDistPoint, tlsCrlDistPoint)
+        val nodeCert = recreateNodeCaAndTlsCertificates(signingCertificateStore, p2pSslConfiguration, nodeCrlDistPoint, tlsCrlDistPoint)
         val keyStore = clientConfig.p2pSslOptions.keyStore.get()
 
         val amqpConfig = object : AMQPConfiguration {
@@ -444,7 +382,7 @@ class CertificateRevocationListNodeTests {
             doReturn(signingCertificateStore).whenever(it).signingCertificateStore
         }
         serverConfig.configureWithDevSSLCertificate()
-        val nodeCert = (signingCertificateStore to p2pSslConfiguration).recreateNodeCaAndTlsCertificates(nodeCrlDistPoint, tlsCrlDistPoint)
+        val nodeCert = recreateNodeCaAndTlsCertificates(signingCertificateStore, p2pSslConfiguration, nodeCrlDistPoint, tlsCrlDistPoint)
         val keyStore = serverConfig.p2pSslOptions.keyStore.get()
         val amqpConfig = object : AMQPConfiguration {
             override val keyStore = keyStore
@@ -457,24 +395,28 @@ class CertificateRevocationListNodeTests {
         return nodeCert
     }
 
-    private fun Pair<CertificateStoreSupplier, MutualSslConfiguration>.recreateNodeCaAndTlsCertificates(nodeCaCrlDistPoint: String?, tlsCrlDistPoint: String?): X509Certificate {
-
-        val signingCertificateStore = first
-        val p2pSslConfiguration = second
+    private fun recreateNodeCaAndTlsCertificates(signingCertificateStore: CertificateStoreSupplier,
+                                                 p2pSslConfiguration: MutualSslConfiguration,
+                                                 nodeCaCrlDistPoint: String?,
+                                                 tlsCrlDistPoint: String?): X509Certificate {
         val nodeKeyStore = signingCertificateStore.get()
         val (nodeCert, nodeKeys) = nodeKeyStore.query { getCertificateAndKeyPair(X509Utilities.CORDA_CLIENT_CA, nodeKeyStore.entryPassword) }
-        val newNodeCert = replaceCrlDistPointCaCertificate(nodeCert, CertificateType.NODE_CA, INTERMEDIATE_CA.keyPair, nodeCaCrlDistPoint)
-        val nodeCertChain = listOf(newNodeCert, INTERMEDIATE_CA.certificate, *nodeKeyStore.query { getCertificateChain(X509Utilities.CORDA_CLIENT_CA) }.drop(2).toTypedArray())
+        val newNodeCert = nodeCert.withCrlDistPoint(crlServer.intermediateCa.keyPair, nodeCaCrlDistPoint)
+        val nodeCertChain = listOf(newNodeCert, crlServer.intermediateCa.certificate) +
+                nodeKeyStore.query { getCertificateChain(X509Utilities.CORDA_CLIENT_CA) }.drop(2)
+
         nodeKeyStore.update {
             internal.deleteEntry(X509Utilities.CORDA_CLIENT_CA)
         }
         nodeKeyStore.update {
             setPrivateKey(X509Utilities.CORDA_CLIENT_CA, nodeKeys.private, nodeCertChain, nodeKeyStore.entryPassword)
         }
+
         val sslKeyStore = p2pSslConfiguration.keyStore.get()
         val (tlsCert, tlsKeys) = sslKeyStore.query { getCertificateAndKeyPair(X509Utilities.CORDA_CLIENT_TLS, sslKeyStore.entryPassword) }
-        val newTlsCert = replaceCrlDistPointCaCertificate(tlsCert, CertificateType.TLS, nodeKeys, tlsCrlDistPoint, X500Name.getInstance(ROOT_CA.certificate.subjectX500Principal.encoded))
-        val sslCertChain = listOf(newTlsCert, newNodeCert, INTERMEDIATE_CA.certificate, *sslKeyStore.query { getCertificateChain(X509Utilities.CORDA_CLIENT_TLS) }.drop(3).toTypedArray())
+        val newTlsCert = tlsCert.withCrlDistPoint(nodeKeys, tlsCrlDistPoint, X500Name.getInstance(crlServer.rootCa.certificate.subjectX500Principal.encoded))
+        val sslCertChain = listOf(newTlsCert, newNodeCert, crlServer.intermediateCa.certificate) +
+                sslKeyStore.query { getCertificateChain(X509Utilities.CORDA_CLIENT_TLS) }.drop(3)
 
         sslKeyStore.update {
             internal.deleteEntry(X509Utilities.CORDA_CLIENT_TLS)
@@ -483,124 +425,6 @@ class CertificateRevocationListNodeTests {
             setPrivateKey(X509Utilities.CORDA_CLIENT_TLS, tlsKeys.private, sslCertChain, sslKeyStore.entryPassword)
         }
         return newNodeCert
-    }
-
-    private fun replaceCrlDistPointCaCertificate(currentCaCert: X509Certificate, certType: CertificateType, issuerKeyPair: KeyPair, crlDistPoint: String?, crlIssuer: X500Name? = null): X509Certificate {
-        val signatureScheme = Crypto.findSignatureScheme(issuerKeyPair.private)
-        val provider = Crypto.findProvider(signatureScheme.providerName)
-        val issuerSigner = ContentSignerBuilder.build(signatureScheme, issuerKeyPair.private, provider)
-        val builder = X509Utilities.createPartialCertificate(
-                certType,
-                currentCaCert.issuerX500Principal,
-                issuerKeyPair.public,
-                currentCaCert.subjectX500Principal,
-                currentCaCert.publicKey,
-                Pair(Date(System.currentTimeMillis() - 5.minutes.toMillis()), Date(System.currentTimeMillis() + 10.days.toMillis())),
-                null
-        )
-        crlDistPoint?.let {
-            val distPointName = DistributionPointName(GeneralNames(GeneralName(GeneralName.uniformResourceIdentifier, it)))
-            val crlIssuerGeneralNames = crlIssuer?.let {
-                GeneralNames(GeneralName(crlIssuer))
-            }
-            val distPoint = DistributionPoint(distPointName, null, crlIssuerGeneralNames)
-            builder.addExtension(Extension.cRLDistributionPoints, false, CRLDistPoint(arrayOf(distPoint)))
-        }
-        return builder.build(issuerSigner).toJca()
-    }
-
-    @Path("crl")
-    inner class CrlServlet(private val server: CrlServer) {
-
-        private val SIGNATURE_ALGORITHM = "SHA256withECDSA"
-
-        @GET
-        @Path("node.crl")
-        @Produces("application/pkcs7-crl")
-        fun getNodeCRL(): Response {
-            return Response.ok(createRevocationList(
-                    server,
-                    SIGNATURE_ALGORITHM,
-                    INTERMEDIATE_CA.certificate,
-                    INTERMEDIATE_CA.keyPair.private,
-                    "node.crl",
-                    false,
-                    revokedNodeCerts
-            ).encoded).build()
-        }
-
-        @GET
-        @Path(FORBIDDEN_CRL)
-        @Produces("application/pkcs7-crl")
-        fun getNodeSlowCRL(): Response {
-            return Response.status(Response.Status.FORBIDDEN).build()
-        }
-
-        @GET
-        @Path("intermediate.crl")
-        @Produces("application/pkcs7-crl")
-        fun getIntermediateCRL(): Response {
-            return Response.ok(createRevocationList(
-                    server,
-                    SIGNATURE_ALGORITHM,
-                    ROOT_CA.certificate,
-                    ROOT_CA.keyPair.private,
-                    "intermediate.crl",
-                    false,
-                    revokedIntermediateCerts
-            ).encoded).build()
-        }
-
-        @GET
-        @Path("empty.crl")
-        @Produces("application/pkcs7-crl")
-        fun getEmptyCRL(): Response {
-            return Response.ok(createRevocationList(
-                    server,
-                    SIGNATURE_ALGORITHM,
-                    ROOT_CA.certificate,
-                    ROOT_CA.keyPair.private,
-                    "empty.crl",
-                    true, emptyList()
-            ).encoded).build()
-        }
-    }
-
-    inner class CrlServer(hostAndPort: NetworkHostAndPort) : Closeable {
-
-        private val server: Server = Server(InetSocketAddress(hostAndPort.host, hostAndPort.port)).apply {
-            handler = HandlerCollection().apply {
-                addHandler(buildServletContextHandler())
-            }
-        }
-
-        val hostAndPort: NetworkHostAndPort
-            get() = server.connectors.mapNotNull { it as? ServerConnector }
-                    .map { NetworkHostAndPort(it.host, it.localPort) }
-                    .first()
-
-        override fun close() {
-            println("Shutting down network management web services...")
-            server.stop()
-            server.join()
-        }
-
-        fun start() {
-            server.start()
-            println("Network management web services started on $hostAndPort")
-        }
-
-        private fun buildServletContextHandler(): ServletContextHandler {
-            val crlServer = this
-            return ServletContextHandler().apply {
-                contextPath = "/"
-                val resourceConfig = ResourceConfig().apply {
-                    register(CrlServlet(crlServer))
-                }
-                val jerseyServlet = ServletHolder(ServletContainer(resourceConfig)).apply { initOrder = 0 }
-                addServlet(jerseyServlet, "/*")
-            }
-        }
     }
 
     private fun verifyAMQPConnection(crlCheckSoftFail: Boolean,
@@ -656,7 +480,7 @@ class CertificateRevocationListNodeTests {
             doReturn(crlCheckArtemisServer).whenever(it).crlCheckArtemisServer
         }
         artemisConfig.configureWithDevSSLCertificate()
-        (signingCertificateStore to p2pSslConfiguration).recreateNodeCaAndTlsCertificates(nodeCrlDistPoint, null)
+        recreateNodeCaAndTlsCertificates(signingCertificateStore, p2pSslConfiguration, nodeCrlDistPoint, null)
 
         val server = ArtemisMessagingServer(artemisConfig, artemisConfig.p2pAddress, MAX_MESSAGE_SIZE, null)
         val client = ArtemisMessagingClient(artemisConfig.p2pSslOptions, artemisConfig.p2pAddress, MAX_MESSAGE_SIZE)
